requirements.txt
# ============================================================================
# Parakeet RunPod Serverless - Python Dependencies
# ============================================================================
# NOTE: This file is for reference only.
# All packages are installed to the virtual environment on the volume during
# the bootstrap process. This file documents the versions and dependencies used.
# ============================================================================

# ---------------------------------------------------------------------------
# Core ML Frameworks
# ---------------------------------------------------------------------------

# PyTorch with CUDA 12.8 support
# Installation: pip install torch==2.8.0 torchvision==0.23.0 torchaudio==2.8.0 --index-url https://download.pytorch.org/whl/cu128
torch==2.8.0
torchvision==0.23.0
torchaudio==2.8.0

# ---------------------------------------------------------------------------
# Flash Attention
# ---------------------------------------------------------------------------

# Flash Attention 2.8.1 (installed from wheel)
# Installation: pip install https://github.com/Dao-AILab/flash-attention/releases/download/v2.8.1/flash_attn-2.8.1+cu12torch2.8cxx11abiTRUE-cp312-cp312-linux_x86_64.whl
# flash_attn==2.8.1

# ---------------------------------------------------------------------------
# NVIDIA NeMo Toolkit
# ---------------------------------------------------------------------------

# NeMo Toolkit with ASR support
# Installation: pip install nemo-toolkit[asr]
nemo-toolkit[asr]

# ---------------------------------------------------------------------------
# Hugging Face Integration
# ---------------------------------------------------------------------------

# Hugging Face Hub CLI (includes the new 'hf' command)
huggingface_hub>=0.25.0

# ---------------------------------------------------------------------------
# Audio Processing
# ---------------------------------------------------------------------------

# FFmpeg Python bindings for audio conversion
ffmpeg-python>=0.2.0

# ---------------------------------------------------------------------------
# S3 Support
# ---------------------------------------------------------------------------

# Boto3 for S3 downloads (optional, for URL-based input)
boto3>=1.34.0

# ---------------------------------------------------------------------------
# RunPod SDK
# ---------------------------------------------------------------------------

# RunPod Serverless SDK
# Note: This is installed in the container (not venv) via Dockerfile
runpod>=1.7.0
